chore: 체스와 관련된 테스트파일을 chess 패키지 밑으로 이동
