+++ src/test/java/chess/domain/position/DirectionTest.java
@@ -1,9 +1,5 @@
-package domain.position;
+package chess.domain.position;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import chess.domain.position.Direction;
-import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,6 +7,8 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("Direction은 ")
 class DirectionTest {
